@@ -3,10 +3,10 @@ description: "A new Flutter project."
 
 publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 
-version: 0.0.1+4
+version: 0.0.1+5
 
 environment:
-  sdk: ^3.5.3
+  sdk: ^3.6.1
 
 dependencies:
   flutter:
@@ -40,11 +40,12 @@ dependencies:
   flutter_html: ^3.0.0-beta.2
   crypto: ^3.0.6
   
-  social_media_recorder: ^1.1.13
+  social_media_recorder: ^1.2.1 
   audio_waveforms: ^1.1.0
   video_player: ^2.9.2
   
   syncfusion_flutter_calendar: ^28.1.33
+  google_maps_flutter: ^2.10.1
 
 dev_dependencies:
   flutter_test:
