@@ -32,39 +32,39 @@ class BusinessEntity {
   }) : inHiveAt = DateTime.now();
 
   @HiveField(0)
-  final String businessID;
+  final String? businessID;
   @HiveField(1)
-  final LocationEntity location;
+  final LocationEntity? location;
   @HiveField(2)
-  final bool acceptPromotions;
+  final bool? acceptPromotions;
   @HiveField(3)
-  final String locationType;
+  final String? locationType;
   @HiveField(4)
-  final BusinessTravelDetailEntity travelDetail;
+  final BusinessTravelDetailEntity? travelDetail;
   @HiveField(5)
-  final List<BusinessEmployeeEntity> employees;
+  final List<BusinessEmployeeEntity>? employees;
   @HiveField(6)
-  final BusinessAddressEntity address;
+  final BusinessAddressEntity? address;
   @HiveField(7)
-  final String displayName;
+  final String? displayName;
   @HiveField(8)
-  final String ownerID;
+  final String? ownerID;
   @HiveField(9)
-  final String tagline;
+  final String? tagline;
   @HiveField(10)
-  final String phoneNumber;
+  final String? phoneNumber;
   @HiveField(11)
-  final String companyNo;
+  final String? companyNo;
   @HiveField(12)
-  final List<RoutineEntity> routine;
+  final List<RoutineEntity>? routine;
   @HiveField(13)
-  final List<double> listOfReviews;
+  final List<double>? listOfReviews;
   @HiveField(14)
-  final DateTime createdAt;
+  final DateTime? createdAt;
   @HiveField(15)
-  final DateTime updatedAt;
+  final DateTime? updatedAt;
   @HiveField(16)
   final AttachmentEntity? logo;
   @HiveField(99)
-  final DateTime inHiveAt;
+  final DateTime? inHiveAt;
 }
