@@ -7,7 +7,6 @@ import 'local/local_auth.dart';
 
 abstract interface class SigninRemoteSource {
   Future<DataState<bool>> signin(String email, String password);
-  Future<DataState<bool>> forgotPassword(String email);
 }
 
 class SigninRemoteSourceImpl implements SigninRemoteSource {
@@ -18,9 +17,10 @@ class SigninRemoteSourceImpl implements SigninRemoteSource {
         endpoint: 'userAuth/login',
         requestType: ApiRequestType.post,
         body: json.encode(<String, String>{
-          'email': email,
-          'password': password,
+          'email': email.trim(),
+          'password': password.trim(),
         }),
+        isConnectType: true,
         isAuth: false,
       );
       if (responce is DataSuccess<bool>) {
@@ -31,7 +31,7 @@ class SigninRemoteSourceImpl implements SigninRemoteSource {
         await LocalAuth().signin(currentUser);
         return responce;
       } else {
-        debugPrint('Signin Faile in Remote Source');
+        debugPrint('Signin Failed in Remote Source');
         return DataFailer<bool>(CustomException('Signin Failed'));
       }
     } catch (e) {
@@ -40,14 +40,4 @@ class SigninRemoteSourceImpl implements SigninRemoteSource {
     }
   }
 
-  @override
-  Future<DataState<bool>> forgotPassword(String email) async {
-    try {
-      // Forgot Password
-    } catch (e) {
-      debugPrint('Forget Password Catch in Remote Source - $e');
-      return DataFailer<bool>(CustomException('Forget Password Failed: $e'));
-    }
-    return DataFailer<bool>(CustomException('Forget Password Failed'));
-  }
 }
