@@ -1,17 +1,17 @@
-import '../../../../../../core/usecase/usecase.dart';
-import '../repositories/signin_repository.dart';
+// import '../../../../../../core/usecase/usecase.dart';
+// import '../repositories/signin_repository.dart';
 
-class ForgotPasswordUsecase implements UseCase<bool, String> {
-  const ForgotPasswordUsecase(this.repository);
-  final SigninRepository repository;
+// // class ForgotPasswordUsecase implements UseCase<bool, String> {
+// //   const ForgotPasswordUsecase(this.repository);
+// //   final SigninRepository repository;
 
-  @override
-  Future<DataState<bool>> call(String params) async {
-    try {
-      return await repository.forgotPassword(params);
-    } catch (e) {
-      print(e);
-    }
-    return DataFailer<bool>(CustomException('Error'));
-  }
-}
+// //   @override
+// //   Future<DataState<bool>> call(String params) async {
+// //     try {
+// //       return await repository.forgotPassword(params);
+// //     } catch (e) {
+// //       print(e);
+// //     }
+// //     return DataFailer<bool>(CustomException('Error'));
+// //   }
+// // }
