@@ -1,4 +1,6 @@
+
 enum SignupPageType {
+  accountType,
   basicInfo,
   otp,
   photoVerification,
@@ -7,6 +9,8 @@ enum SignupPageType {
 
   SignupPageType? next() {
     switch (this) {
+      case SignupPageType.accountType:
+        return SignupPageType.basicInfo;
       case SignupPageType.basicInfo:
         return SignupPageType.otp;
       case SignupPageType.otp:
@@ -22,8 +26,10 @@ enum SignupPageType {
 
   SignupPageType? previous() {
     switch (this) {
-      case SignupPageType.basicInfo:
+      case SignupPageType.accountType:
         return null;
+      case SignupPageType.basicInfo:
+        return SignupPageType.accountType;
       case SignupPageType.otp:
         return SignupPageType.basicInfo;
       case SignupPageType.photoVerification:
