@@ -12,104 +12,124 @@ import '../../../domain/entities/listing_entity.dart';
 import '../../../domain/entities/sub_category_entity.dart';
 import 'category_selection_bottom_sheet.dart';
 
-class SubCategorySelectableWidget extends StatelessWidget {
+class SubCategorySelectableWidget extends StatefulWidget {
   const SubCategorySelectableWidget({
     required this.listType,
     required this.subCategory,
     required this.onSelected,
     super.key,
   });
+
   final ListingType? listType;
   final SubCategoryEntity? subCategory;
   final void Function(SubCategoryEntity?) onSelected;
 
   @override
+  State<SubCategorySelectableWidget> createState() =>
+      _SubCategorySelectableWidgetState();
+}
+
+class _SubCategorySelectableWidgetState
+    extends State<SubCategorySelectableWidget> {
+  late Future<List<ListingEntity>> _listingsFuture;
+
+  @override
+  void initState() {
+    super.initState();
+    // Initialize future ONCE to prevent re-triggering
+    _listingsFuture = ListingAPI().listing();
+  }
+
+  @override
   Widget build(BuildContext context) {
     return FutureBuilder<List<ListingEntity>>(
-        future: ListingAPI().listing(),
-        // initialData: LocalListing().listings,
-        builder: (
-          BuildContext context,
-          AsyncSnapshot<List<ListingEntity>> snapshot,
-        ) {
-          if (snapshot.connectionState == ConnectionState.waiting &&
-              kDebugMode) {
-            return const Loader();
-          }
-          if (snapshot.hasError && snapshot.data == null) {
-            AppLog.error(
-              'SubCategorySelectableWidget: snapshot -> ${snapshot.error}',
-              name: 'SubCategorySelectableWidget.build - snapshot',
-              error: snapshot.error,
-            );
-            return Center(child: Text('something_wrong'.tr()));
-          }
-          AppLog.info(
-            'SubCategorySelectableWidget: snapshot -> ${snapshot.data?.length}',
-            name: 'SubCategorySelectableWidget.build - snapshot',
+      future: _listingsFuture,
+      builder: (
+        BuildContext context,
+        AsyncSnapshot<List<ListingEntity>> snapshot,
+      ) {
+        // Early return for loading/error states
+        if (snapshot.connectionState == ConnectionState.waiting) {
+          return const Loader();
+        }
+
+        if (snapshot.hasError || !snapshot.hasData) {
+          AppLog.error(
+            'SubCategorySelectableWidget: ${snapshot.error}',
+            name: 'SubCategorySelectableWidget.build',
+            error: snapshot.error,
           );
-          return Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: <Widget>[
-              const Text(
-                'category',
-                style: TextStyle(fontWeight: FontWeight.w500),
-              ).tr(),
-              const SizedBox(height: 4),
-              InkWell(
-                onTap: () async {
-                  final List<ListingEntity> selectedList = snapshot.data
-                          ?.where((ListingEntity element) =>
-                              element.type == listType || kDebugMode)
-                          .toList() ??
-                      <ListingEntity>[];
-                  if (selectedList.isEmpty) {
-                    AppSnackBar.showSnackBar(context, 'something_wrong'.tr());
-                    return;
-                  }
-                  final SubCategoryEntity? selectCat =
-                      await showModalBottomSheet(
-                    context: context,
-                    builder: (BuildContext context) {
-                      return CategorySelectionBottomSheet(
-                        subCategories: selectedList.first.subCategory,
-                      );
-                    },
-                  );
-                  if (selectCat != null) {
-                    onSelected(selectCat);
-                  }
-                },
-                borderRadius: BorderRadius.circular(8),
-                child: Container(
-                  width: double.infinity,
-                  padding:
-                      const EdgeInsets.symmetric(horizontal: 16, vertical: 8),
-                  decoration: BoxDecoration(
-                    borderRadius: BorderRadius.circular(8),
-                    border: Border.all(color: Colors.grey),
-                  ),
-                  child: Row(
-                    children: <Widget>[
-                      subCategory == null
-                          ? Text(
-                              subCategory?.title ?? 'select_sub_category'.tr(),
-                              maxLines: 1,
-                              overflow: TextOverflow.ellipsis,
-                              style: const TextStyle(
-                                fontSize: 16,
-                                color: Colors.grey,
-                              ),
-                            )
-                          : Text(subCategory?.title ?? 'Null Title'),
-                      const Spacer(),
-                      const Icon(Icons.arrow_drop_down, color: Colors.grey),
-                    ],
+          return Center(child: Text('something_wrong'.tr()));
+        }
+
+        // Cache filtered data to avoid recalculating
+        final List<ListingEntity> selectedList = snapshot.data!
+            .where((ListingEntity element) =>
+                element.type == widget.listType || kDebugMode)
+            .toList();
+
+        return _buildCategorySelector(selectedList);
+      },
+    );
+  }
+
+  Widget _buildCategorySelector(List<ListingEntity> selectedList) {
+    return Column(
+      crossAxisAlignment: CrossAxisAlignment.start,
+      children: <Widget>[
+        const Text(
+          'category',
+          style: TextStyle(fontWeight: FontWeight.w500),
+        ).tr(),
+        const SizedBox(height: 4),
+        InkWell(
+          onTap: () => _handleCategorySelection(selectedList),
+          borderRadius: BorderRadius.circular(8),
+          child: Container(
+            width: double.infinity,
+            padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 8),
+            decoration: BoxDecoration(
+              borderRadius: BorderRadius.circular(8),
+              border: Border.all(color: Colors.grey),
+            ),
+            child: Row(
+              children: <Widget>[
+                Expanded(
+                  child: Text(
+                    widget.subCategory?.title ?? 'select_sub_category'.tr(),
+                    maxLines: 1,
+                    overflow: TextOverflow.ellipsis,
+                    style: TextStyle(
+                      fontSize: 16,
+                      color: widget.subCategory == null ? Colors.grey : null,
+                    ),
                   ),
                 ),
-              ),
-            ],
-          );
-        });
+                const Icon(Icons.arrow_drop_down, color: Colors.grey),
+              ],
+            ),
+          ),
+        ),
+      ],
+    );
+  }
+
+  Future<void> _handleCategorySelection(
+      List<ListingEntity> selectedList) async {
+    if (selectedList.isEmpty) {
+      AppSnackBar.showSnackBar(context, 'something_wrong'.tr());
+      return;
+    }
+
+    final SubCategoryEntity? selectCat = await showModalBottomSheet(
+      context: context,
+      builder: (BuildContext context) => CategorySelectionBottomSheet(
+        subCategories: selectedList.first.subCategory,
+      ),
+    );
+
+    if (selectCat != null) {
+      widget.onSelected(selectCat);
+    }
   }
 }
