@@ -1,7 +1,10 @@
 import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart';
+import 'package:provider/provider.dart';
 
 import '../../../../../../../../../../../core/widgets/custom_elevated_button.dart';
+import '../../../../../../../../../book_visit/view/provider/view_booking_provider.dart';
+import '../../../../../../../../../book_visit/view/screens/view_booking_screen.dart';
 import '../../../../../../../../domain/entities/post_entity.dart';
 import 'widgets/post_make_offer_button.dart';
 
@@ -11,6 +14,7 @@ class PostVehicleButtonTile extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
+    final BookingProvider pro = Provider.of<BookingProvider>(context, listen: false);
     return Row(
       children: <Widget>[
         Expanded(child: PostMakeOfferButton(post: post)),
@@ -26,7 +30,11 @@ class PostVehicleButtonTile extends StatelessWidget {
               color: Theme.of(context).primaryColor,
             ),
             isLoading: false,
-            onTap: () {},
+            onTap: () {
+              pro.disposed();
+              Navigator.pushNamed(context, BookingScreen.routeName,
+                  arguments: <String, dynamic>{'post': post});
+            },
           ),
         ),
       ],
