@@ -1,12 +1,14 @@
 import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-
 import '../../../../../../core/sources/data_state.dart';
 import '../../../../../../core/widgets/attachment_slider.dart';
+import '../../../../auth/signin/data/sources/local/local_auth.dart';
 import '../../../data/sources/local/local_post.dart';
 import '../../../domain/entities/post_entity.dart';
+import '../../../domain/entities/visit/visiting_entity.dart';
 import '../../../feed/views/widgets/post/widgets/section/buttons/home_post_button_section.dart';
+import '../../../feed/views/widgets/post/widgets/section/buttons/type/post_button_for_user_tile.dart';
 import '../../../feed/views/widgets/post/widgets/section/home_post_header_section.dart';
 import '../providers/post_detail_provider.dart';
 import '../widgets/post_detail_description_section.dart';
@@ -27,6 +29,7 @@ class PostDetailScreen extends StatelessWidget {
     final Map<String, dynamic> args =
         ModalRoute.of(context)!.settings.arguments as Map<String, dynamic>;
     final String postID = args['pid'] ?? '';
+    final VisitingEntity? visit = args['visit'] as VisitingEntity?;
     return Scaffold(
       appBar: AppBar(title: const Text('post_details').tr()),
       body: FutureBuilder<DataState<PostEntity>>(
@@ -39,6 +42,7 @@ class PostDetailScreen extends StatelessWidget {
         ) {
           final PostEntity? post =
               snapshot.data?.entity ?? LocalPost().post(postID);
+          final bool isMe = post?.createdBy == (LocalAuth.uid ?? '-');
           return post == null
               ? const SizedBox()
               : Padding(
@@ -54,10 +58,13 @@ class PostDetailScreen extends StatelessWidget {
                         AttachmentsSlider(urls: post.fileUrls),
                         PostDetailTitleAmountSection(post: post),
                         PostRatingSection(post: post),
-                        PostButtonSection(
-                          post: post,
-                          padding: const EdgeInsets.symmetric(vertical: 8),
-                        ),
+                        if (isMe == true)
+                          PostButtonsForUser(visit: visit, post: post),
+                        if (isMe == false)
+                          PostButtonSection(
+                            post: post,
+                            padding: const EdgeInsets.symmetric(vertical: 8),
+                          ),
                         PostDetailDescriptionSection(post: post),
                         PostDetailTileListSection(post: post),
                         PostDetailPostageReturnSection(post: post),
