@@ -4,7 +4,9 @@ import '../../../../../core/functions/app_log.dart';
 import '../../../../../core/sources/api_call.dart';
 import '../../../../business/business_page/domain/params/get_business_bookings_params.dart';
 import '../../../../business/business_page/domain/usecase/get_business_bookings_list_usecase.dart';
+import '../../../../business/core/domain/entity/business_entity.dart';
 import '../../../../business/core/domain/entity/service/service_entity.dart';
+import '../../../../business/core/domain/usecase/get_business_by_id_usecase.dart';
 import '../../../auth/signin/data/sources/local/local_auth.dart';
 import '../../../bookings/data/sources/local_booking.dart';
 import '../../../bookings/domain/entity/booking_entity.dart';
@@ -16,10 +18,12 @@ import '../enums/services_page_type.dart';
 class ServicesPageProvider extends ChangeNotifier {
   ServicesPageProvider(
     this._getSpecialOfferUsecase,
-    this._getBookingsListUsecase,
+    this._getBookingsListUsecase, this._getBusinessByIdUsecase,
+    
   );
   final GetSpecialOfferUsecase _getSpecialOfferUsecase;
   final GetBookingsListUsecase _getBookingsListUsecase;
+  final GetBusinessByIdUsecase _getBusinessByIdUsecase;
   //
   Future<List<ServiceEntity>> getSpecialOffer() async {
     try {
@@ -70,6 +74,11 @@ class ServicesPageProvider extends ChangeNotifier {
       );
     }
     return LocalBooking().userBooking(LocalAuth.uid ?? '');
+  }
+
+
+  Future<DataState<BusinessEntity?>> getbusinessbyid(String uid) async {
+    return await _getBusinessByIdUsecase(uid);
   }
 
   //
