@@ -1,16 +1,30 @@
 import 'package:flutter/material.dart';
-
 import '../core/functions/app_log.dart';
 import '../features/business/service/views/screens/add_service_screen.dart';
+import '../features/personal/auth/find_account/view/screens/confirm_email_screen.dart';
+import '../features/personal/auth/find_account/view/screens/enter_code_screen.dart';
+import '../features/personal/auth/find_account/view/screens/find_account_screen.dart';
+import '../features/personal/auth/find_account/view/screens/new_password_screen.dart';
+import '../features/personal/auth/find_account/view/screens/send_code_screen.dart';
 import '../features/personal/cart/views/screens/personal_cart_screen.dart';
 import '../features/personal/chats/chat/views/screens/chat_screen.dart';
+import '../features/personal/explore/views/screens/filter_categories/explore_cloth_foot_screen.dart';
+import '../features/personal/explore/views/screens/filter_categories/explore_food_drink_screen.dart';
+import '../features/personal/explore/views/screens/filter_categories/explore_pets_screen.dart';
+import '../features/personal/explore/views/screens/filter_categories/explore_popular_screen.dart';
+import '../features/personal/explore/views/screens/filter_categories/explore_property_screen.dart';
+import '../features/personal/explore/views/screens/filter_categories/explore_vehicles_screen.dart';
 import '../features/personal/listing/listing_form/views/screens/add_listing_form_screen.dart';
 import '../features/personal/auth/signin/views/screens/sign_in_screen.dart';
 import '../features/personal/auth/signup/views/screens/signup_screen.dart';
 import '../features/personal/dashboard/views/screens/dasboard_screen.dart';
+import '../features/personal/book_visit/view/screens/view_booking_screen.dart';
 import '../features/personal/post/post_detail/views/screens/post_detail_screen.dart';
+import '../features/personal/review/features/reivew_list/views/screens/media_picker_screen.dart';
+import '../features/personal/review/features/reivew_list/views/screens/write_review_screen.dart';
 import '../features/personal/setting/more_info/views/screens/personal_more_information_setting_screen.dart';
 import '../features/personal/setting/setting_dashboard/views/screens/personal_setting_screen.dart';
+import '../features/personal/user/profiles/views/screens/edit_profile_screen.dart';
 
 class AppRoutes {
   static const String baseURL = 'https://selloutweb.com';
@@ -45,14 +59,21 @@ class AppRoutes {
     // AUTH
     SignInScreen.routeName: (_) => const SignInScreen(),
     SignupScreen.routeName: (_) => const SignupScreen(),
+    FindAccountScreen.routeName: (_) => const FindAccountScreen(),
+    ConfirmEmailScreen.routeName: (_) => const ConfirmEmailScreen(),
+    SendCodeScreen.routeName: (_) => const SendCodeScreen(),
+    EnterCodeScreen.routeName: (_) => const EnterCodeScreen(),
+    NewPasswordScreen.routeName: (_) => const NewPasswordScreen(),
 
     // DAHSBOARD
     DashboardScreen.routeName: (_) => const DashboardScreen(),
     AddListingFormScreen.routeName: (_) => const AddListingFormScreen(),
     // POST
     PostDetailScreen.routeName: (_) => const PostDetailScreen(),
-
+    //CHAT
     ChatScreen.routeName: (_) => const ChatScreen(),
+    //BOOKING
+    BookingScreen.routeName: (_) => const BookingScreen(),
 
     // CART
     PersonalCartScreen.routeName: (_) => const PersonalCartScreen(),
@@ -60,8 +81,19 @@ class AppRoutes {
     PersonalSettingScreen.routeName: (_) => const PersonalSettingScreen(),
     PersonalSettingMoreInformationScreen.routeName: (_) =>
         const PersonalSettingMoreInformationScreen(),
-
+    //Review
+    WriteReviewScreen.routeName: (_) => const WriteReviewScreen(),
+    MediaPickerScreen.routeName: (_) => const MediaPickerScreen(),
     // BUSSINESS
     AddServiceScreen.routeName: (_) => const AddServiceScreen(),
+    // Explore
+    ExplorePopularScreen.routeName: (_) => const ExplorePopularScreen(),
+    ExploreCLothFOotScreen.routeName: (_) => const ExploreCLothFOotScreen(),
+    ExplorePetsScreen.routeName: (_) => const ExplorePetsScreen(),
+    ExplorePropertyScreen.routeName: (_) => const ExplorePropertyScreen(),
+    ExploreFoodDrinkScreen.routeName: (_) => const ExploreFoodDrinkScreen(),
+    ExploreVehiclesScreen.routeName: (_) => const ExploreVehiclesScreen(),
+    // Profile
+    EditProfileScreen.routeName: (_) => const EditProfileScreen(),
   };
 }
